add children tests
